requirements.txt
streamlit==1.29.0
yfinance==0.2.18
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
ta==0.10.2
textblob==0.17.1
vaderSentiment==3.3.2
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
mplfinance==0.12.10b0